+++ plugins/module_utils/entity.py
@@ -2,14 +2,19 @@
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 from __future__ import absolute_import, division, print_function
 
+from http.client import IncompleteRead
+
 __metaclass__ = type
 
 import copy
 import json
+import os
 from base64 import b64encode
 
 from ansible.module_utils._text import to_text
 from ansible.module_utils.urls import fetch_url
+
+from ..module_utils import utils
 
 try:
     from urllib.parse import parse_qsl, urlencode, urlparse, urlunparse
@@ -18,6 +23,9 @@ except ImportError:
 
 
 class Entity(object):
+    entities_limitation = 250
+    entity_type = "entities"
+
     def __init__(
         self,
         module,
@@ -98,6 +106,28 @@ class Entity(object):
             timeout=timeout,
         )
 
+    # source is the file path of resource where ansible yaml runs
+    def upload(
+        self,
+        source,
+        endpoint=None,
+        query=None,
+        raise_error=True,
+        no_response=False,
+        timeout=30,
+    ):
+        url = self.base_url + "/{0}".format(endpoint) if endpoint else self.base_url
+        if query:
+            url = self._build_url_with_query(url, query)
+        return self._upload_file(
+            url,
+            source,
+            method="POST",
+            raise_error=raise_error,
+            no_response=no_response,
+            timeout=timeout,
+        )
+
     def delete(
         self,
         uuid=None,
@@ -132,14 +162,32 @@ class Entity(object):
         url = self.base_url if use_base_url else self.base_url + "/list"
         if endpoint:
             url = url + "/{0}".format(endpoint)
-        return self._fetch_url(
-            url,
-            method="POST",
-            data=data,
-            raise_error=raise_error,
-            no_response=no_response,
-            timeout=timeout,
-        )
+        entities_list = []
+        while True:
+            resp = self._fetch_url(
+                url,
+                method="POST",
+                data=data,
+                raise_error=raise_error,
+                no_response=no_response,
+                timeout=timeout,
+            )
+            if not resp.get(self.entity_type):
+                return resp
+            entities_list.extend(resp[self.entity_type])
+            entities_count = len(entities_list)
+            data["offset"] = entities_count
+            if len(resp[self.entity_type]) != self.entities_limitation:
+                break
+        custom_filters = self.module.params.get("custom_filter")
+        if custom_filters:
+            entities_list = self.filter_entities(entities_list, custom_filters)
+            entities_count = len(entities_list)
+
+        resp[self.entity_type] = entities_list
+        resp["metadata"]["length"] = entities_count
+        resp["metadata"]["total_matches"] = entities_count
+        return resp
 
     def get_spec(self, old_spec=None):
         spec = copy.deepcopy(old_spec) or self._get_default_spec()
@@ -206,11 +254,65 @@ class Entity(object):
     def _fetch_url(
         self, url, method, data=None, raise_error=True, no_response=False, timeout=30
     ):
-        data = self.module.jsonify(data) if data else None
+
+        # only jsonify if content-type supports, added to avoid incase of form-url-encodeded type data
+        if self.headers["Content-Type"] == "application/json":
+            data = self.module.jsonify(data) if data else None
+
         resp, info = fetch_url(
             self.module,
             url,
             data=data,
+            method=method,
+            headers=self.headers,
+            cookies=self.cookies,
+            timeout=timeout,
+        )
+
+        status_code = info.get("status")
+        body = resp.read() if resp else info.get("body")
+        while True:
+            try:
+                resp_json = json.loads(to_text(body)) if body else None
+            except ValueError:
+                resp_json = None
+            except IncompleteRead:
+                continue
+
+            if not raise_error:
+                return resp_json
+
+            if status_code >= 300:
+                err = info.get("msg", "Status code != 2xx")
+                self.module.fail_json(
+                    msg="Failed fetching URL: {0}".format(url),
+                    status_code=status_code,
+                    error=err,
+                    response=resp_json,
+                )
+
+            if no_response:
+                return {"status_code": status_code}
+
+            if not resp_json:
+                self.module.fail_json(
+                    msg="Failed to convert API response to json",
+                    status_code=status_code,
+                    error=body,
+                    response=resp_json,
+                )
+
+            return resp_json
+
+    # upload file in chunks to the given url
+    def _upload_file(
+        self, url, source, method, raise_error=True, no_response=False, timeout=30
+    ):
+
+        resp, info = fetch_url(
+            self.module,
+            url,
+            data=FileChunksIterator(source),
             method=method,
             headers=self.headers,
             cookies=self.cookies,
@@ -248,3 +350,61 @@ class Entity(object):
             )
 
         return resp_json
+
+    def unify_spec(self, spec1, spec2):
+        """
+        This routine return intersection of two specs(dict) as per
+        keys in first level of dictionary.
+        """
+        spec = {}
+        for k in spec1:
+            v = spec2.get(k)
+            if v:
+                spec[k] = v
+        return spec
+
+    @staticmethod
+    def parse_filters(filters):
+        return ",".join(map(lambda i: "{0}=={1}".format(i[0], i[1]), filters.items()))
+
+    @staticmethod
+    def filter_entities(entities, custom_filters):
+        filtered_entities = []
+        for entity in entities:
+            if utils.intersection(entity, custom_filters.copy()):
+                filtered_entities.append(entity)
+        return filtered_entities
+
+
+# Read files in chunks and yeild it
+class CreateChunks(object):
+    def __init__(self, filename, chunk_size=1 << 13):
+        self.filename = filename
+        self.chunk_size = chunk_size
+        self.total_size = os.path.getsize(filename)
+
+    def __iter__(self):
+        with open(self.filename, "rb") as file:
+            while True:
+                data = file.read(self.chunk_size)
+                if not data:
+                    break
+                yield data
+
+    def __len__(self):
+        return self.total_size
+
+
+# to iterate over chunks of file
+class FileChunksIterator(object):
+    def __init__(self, filename, chunk_size=1 << 13):
+        iterable = CreateChunks(filename, chunk_size)
+        self.iterator = iter(iterable)
+        self.length = len(iterable)
+
+    # request lib checks for read func in iterable object
+    def read(self, size=None):
+        return next(self.iterator, b"")
+
+    def __len__(self):
+        return self.length